backupccl: add more fine-grained privilege checks to BACKUP
This changes removes the umbrella admin check from the BACKUP
code path and introduces more role based privilege checking.
This is in keeping with the changes which have recently been made
to IMPORT and RESTORE.

BACKUPS now have the following privilege checks:
- Cluster BACKUPs can only be performed by admin users.
- Tenant BACKUP can only be performed by admin users.
- ExternalStorage URIs which require an admin access
can only be used as a BACKUP source if the user is an admin.
- Database and table BACKUP require SELECT privileges on each respectively.
- UDT, UDS require USAGE privileges on each respectively.

Fixes: #54635

Release note: None