rowexec: optimize partially ordered distinct
This commit optimizes the partially ordered distinct by fixing a bug in
the implementation of the optimization's intention. When we have
partially ordered input stream, it is sufficient to only hash
non-ordered columns in order to distinguish between the rows. That has
also been the intention, but we actually hashed all distinct columns.
This is now fixed which gives a minor speedup.

Release note: None