cli/sql: implement the `\i` and `\ir` client-side command
Release note (cli change): The SQL shell (`cockroach sql`, `cockroach
demo`) now support the `\i` and `\ir` client-side command which reads
SQL file and evaluates its content in-place.

`\ir` differs from `\i` in that the file name is resolved relative
to the location of the script containing the `\ir` command. This makes
`\ir` likely more desirable in the general case.

Occurences of `\q` inside a file included via `\i`/`\ir stop
evaluation of the file and resume evaluation of the file that included
it.

This feature is compatible with the identically named `psql` commands.
It is meant to help compose complex initialization scripts from a
library of standard components.

For example, one could be defining each table and its initial contents
in separate SQL files, and then use different super-files to include
different tables depending on the desired final schema.