sql: emit default fields with crdb_internal.pb_to_json
This commit updates `crdb_internal.pb_to_json` to render fields with
zero values. This makes the function more usable with enums or other
field types that are often set to the zero value.

The commit plumbs the `emitDefaults` option up through the protoreflect
package so that users can opt-in to the behavior. It doesn't go as far
as creating an overload for crdb_internal.pb_to_json because it's too
late in the release cycle for such a change. Doing so will be trivial
at any future point.