sql: fix error message refering to session variable as cluster setting
This commit updates the error message text which referred to
`experimental_enable_hash_sharded_indexes` as a cluster setting when it
is in fact a session variable.

Release note (bug fix): Error message previously refered to
experimental_enable_hash_sharded_indexes as a cluster setting when it is
in fact a session variable.