backfill: clear bound account after index KV has been written
Previously, we were not clearing the bound account associated
with the index backfiller, once the index KVs (which it was
monitoring memory for) had been written and didn't need to
be referenced in the future.

This change clears the bound account after each "chunk" is
processed by the index backfiller. Since the bound account is
only used to track the memory of the index KVs generated when
processing a "chunk", we can simply Clear() it instead of doing
any fine grained shrinking.

Release note: None