opt: fix nil-pointer error when building invertedFilter
This commit fixes a nil-pointer error that could occur if no
PreFiltererState was set when building an InvertedFilterExpr in the
execbuilder.

Fixes #54773

Release note (bug fix): Fixed a nil pointer error that could occur
at planning time for some spatial queries when an inverted index
existed on a geometry or geography column.