rowexec: remove recently added memory accounting in sampler
We recently added the memory accounting for `tree.Datum`s that could be
allocated during stats collection during fingerprinting. The idea was
that Go GC is slow, so it's a good idea to account for those allocations
even though the references to the datums are lost right away. After
thinking about it more and doing some experiments, I think it's a bad
idea because it could lead to stats jobs hitting a memory budget error
in cases when it shouldn't happen which would result in auto stats never
completing. We will just keep our old behavior of possibly running into
OOMs until we implement the vectorized stats collection and can make
easier improvements with memory accounting.

Release note: None