kvserver: add metric to track failed attempts to close timestamps
This commit adds a store-level metric to track failed attempts to close
timestamps by the min prop tracker, due to either an epoch mismatch or pending
evaluation(s) below the timestamp it attempted to close.

This is intended to aid our debugging of stuck or lagging closed timestamps.

Release note: None