roachtest: de-flake (again again!) roachtest/inconsistent
The consistency check could still happen at a moment at which
not all nodes were live "to each other". A consistency check
that skips a node is between two nodes, so a 50:50 chance of
killing either. Not what we want - we want n1 to be the designated
loser.

Fixes #55205.

Release note: None