workload/schemachange: remove error handling overhead
Previously, certain ops screened for errors while others
did not. For this reason, there was some overhead in terms
of states and storage to deal with errors differently depending
on the op that produced the error.

Now, all ops screen for errors, so extra states and mechanisms can
be removed.

Release note: None