workload/schemachange: use regex to fetch tables
Previously, the clause `LIKE 'table%'` was used to fetch tables.
This pattern was capabable of fetching system tables, such as
information_schema.tables. Generally, this workload should not
alter the system tables, and should only test using the tables it
creates. Tables created by the workload can be fetched using
the clause `~ 'table[0-9]+'`, which is the word 'table' followed
by a sequence number.

Release note: None